Refactor dumpgenerator into sub-packages
