+++ wikiteam3/dumpgenerator/api/__init__.py
@@ -1,0 +1,4 @@
+from .api import checkAPI, checkRetryAPI, mwGetAPIAndIndex
+from .get_json import getJSON
+from .handle_status_code import handleStatusCode
+from .wiki_check import getWikiEngine
