+++ wikiteam3/dumpgenerator/cli/__init__.py
@@ -1,0 +1,3 @@
+from .cli import getParameters
+from .greeter import bye, welcome
+from .delay import Delay
