+++ wikiteam3/dumpgenerator/dump/__init__.py
@@ -1,0 +1,1 @@
+from .generator import DumpGenerator
