+++ wikiteam3/dumpgenerator/dump/xmlrev/xml_truncate.py
