+++ wikiteam3/dumpgenerator/log/__init__.py
@@ -1,0 +1,1 @@
+from .log_error import logerror
