+++ wikiteam3/utils/__init__.py
@@ -1,0 +1,7 @@
+from .uprint import uprint
+from .util import removeIP, cleanXML, cleanHTML, undoHTMLEntities
+
+from .user_agent import getUserAgent
+from .domain import domain2prefix
+from .truncate import truncateFilename
+from .wiki_avoid import avoidWikimediaProjects
